+++ brewtils/__init__.py
@@ -8,6 +8,7 @@ from yapconf.exceptions import YapconfItemNotFound
 
 from brewtils.decorators import command, parameter, system
 from brewtils.errors import ValidationError
+from brewtils.log import configure_logging
 from brewtils.plugin import Plugin, RemotePlugin
 from brewtils.rest import normalize_url_prefix
 from brewtils.rest.system_client import SystemClient
@@ -26,6 +27,7 @@ __all__ = [
     'get_connection_info',
     'load_config',
     'get_bg_connection_parameters',
+    'configure_logging',
 ]
 
 __version__ = generated_version